src/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
tortoise-orm==0.19.3
# databases[postgresql]==0.2.6
python-redis-cache==1.2.0
redis==4.5.4
# pika==1.3.1
coloredlogs==15.0.1

# Requests library for making HTTP calls to APIs
requests==2.32.3

# Python dotenv for managing environment variables
python-dotenv==1.0.1

behave==1.2.6

typer==0.15.1